Re-adjustment
Several times I experienced trial and error and the worst thing was that I found a double story "a" that was very similar to what I was doing, this was really an accident. Then I decided to work offline. Changed the Arpertur style and finally found a dry formula for Balinese scrip.

Here are the changes:
1. Readjust aperture Latin became more Grotesk.
2. Readjust the side bearing Latin
3. Generate all diacritics and multilingual Latin.
4. Generate an OpenType feature for Latin.
5. Kerning for Latin
6. And currently, the Balinese script is in the process of kerning and re-adjustment of some characters.